+++ Editor/SequenceAssemblyWindow/SequenceAssemblyPlayModeInspector.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using UnityEditor.Graphs;
 using UnityEngine;
 using UnityEngine.UIElements;
 
@@ -7,30 +6,19 @@ namespace UnityEditor.Sequences
 {
     internal class SequenceAssemblyPlayModeInspector : SequenceAssemblyInspector
     {
-        void OnEnable()
+        internal override void Initialize()
         {
-            Initialize();
-        }
+            base.Initialize();
 
-        void OnDisable()
-        {
-            ClearCollectionsCache();
-        }
-
-        public override VisualElement CreateInspectorGUI()
-        {
-            var root = base.CreateInspectorGUI();
             var playmodeMessage = new Label("Read-only in Play mode.");
             playmodeMessage.AddToClassList("seq-warning-label");
-            root.Insert(0, playmodeMessage);
-
-            return root;
+            rootVisualElement.Insert(0, playmodeMessage);
         }
 
         protected override void SetAssetCollection(string type, IEnumerable<GameObject> sequenceAssetSelections)
         {
-            AssetCollectionList newAssetCollectionList = new AssetCollectionList(m_Director, type, true);
-            AddCollectionList(sequenceAssetSelections, newAssetCollectionList);
+            var assetCollectionList = AssetCollectionList.Get(m_Director, type, true);
+            AddCollectionList(sequenceAssetSelections, assetCollectionList);
         }
 
         protected override void  AddSequenceAssetToCollection(IEnumerable<GameObject> sequenceAssetSelections, AssetCollectionList assetCollectionList)